fix: rename blog.ts to blog.server.ts and lazy-load MDX components (#59)
- Rename app/utils/blog.ts → app/utils/blog.server.ts to enforce the
  server-only boundary and align with docs/facets/blog.md specification
- Update imports in blog/index.tsx, rss[.]xml.ts, and sitemap[.]xml.ts
- Fix $slug/index.tsx: remove getPost() call from component body (which
  was pulling all eagerly-loaded MDX content into the client bundle);
  replace with a non-eager import.meta.glob + React.lazy + Suspense so
  only the single requested post's MDX is loaded on the client
